+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Toast overflows the bottom-right conner, making the app unresponsive in that section.
+
 ## [4.44.2] - 2021-07-12
 
 ### Fixed