+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [4.44.3] - 2021-07-27
+
 ### Changed
 
 - Image preview resolution in the uploader widget.